requirements.txt
#
# Python PIP requirements for degirum_tools external dependencies installation
#
# Copyright 2024 DeGirum Corporation
#

numpy
scipy
pillow
python-dotenv
requests
psutil
opencv-python
degirum>=0.16.2
ipython
pafy
youtube-dl==2020.12.2
faster-coco-eval
pycocotools
pyyaml
ffmpegcv>=0.3.15;platform_system!='Windows'
ffmpeg-python
typing-extensions
jsonschema
apprise
